# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Items Management API",
        "description": "API for managing products, orders, and user authentication"
    },
    "host": "item-management-api.onrender.com",
    "schemes": ["https"],
    "paths": {
        "/auth/register": {
            "post": {
                "summary": "Register a new user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": { "type": "string" },
                                "email": { "type": "string" },
                                "password": { "type": "string" }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": { "description": "User registered successfully" },
                    "400": { "description": "User already exists or validation error" },
                    "500": { "description": "Internal server error" }
                }
            }
        },
        "/auth/register-admin": {
        "post": {
            "summary": "Register an Admin (Restricted)",
            "description": "Allows an admin user to create another admin.",
            "parameters": [
                {
                    "name": "body",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "username": { "type": "string" },
                            "email": { "type": "string" },
                            "password": { "type": "string" }
                        }
                    }
                }
            ],
            "responses": {
                "201": { "description": "Admin registered successfully" },
                "400": { "description": "User already exists" },
                "403": { "description": "Forbidden: Only admins can create admins" },
                "500": { "description": "Internal server error" }
            }
        }
    },
        "/auth/login": {
            "post": {
                "summary": "Log in with username and password",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": { "type": "string" },
                                "password": { "type": "string" }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "Login successful" },
                    "401": { "description": "Unauthorized - invalid username or password" },
                    "500": { "description": "Internal server error" }
                }
            }
        },
        "/auth/github": {
            "get": {
                "summary": "GitHub OAuth Login",
                "description": "Redirects to GitHub for OAuth authentication.",
                "responses": {
                    "302": { "description": "Redirect to GitHub login page" },
                    "500": { "description": "Internal server error" }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "summary": "Log out the current user",
                "description": "Ends the current user's session and logs them out.",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Logout successful"
                                }
                            }
                        }
                    },
                    "500": { "description": "Internal server error" }
                }
            }
        },
        "/api/products": {
            "get": {
                "summary": "Get all products",
                "description": "Retrieve a list of all products",
                "responses": {
                    "200": {
                        "description": "List of products",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Product"
                            }
                        }
                    },
                    "500": { "description": "Server error" }
                }
            },
            "post": {
                "summary": "Create a new product",
                "description": "Add a new product to the database",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    }
                ],
                "responses": {
                    "201": { "description": "Product created successfully" },
                    "400": { "description": "Invalid input data" },
                    "500": { "description": "Server error" }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "summary": "Get a product by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "ID of the product"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    },
                    "404": { "description": "Product not found" },
                    "500": { "description": "Server error" }
                }
            },
            "put": {
                "summary": "Update a product by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "ID of the product to update"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "Product updated successfully" },
                    "400": { "description": "Invalid input data" },
                    "404": { "description": "Product not found" },
                    "500": { "description": "Server error" }
                }
            },
            "delete": {
                "summary": "Delete a product by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "ID of the product to delete"
                    }
                ],
                "responses": {
                    "200": { "description": "Product deleted successfully" },
                    "404": { "description": "Product not found" },
                    "500": { "description": "Server error" }
                }
            }
        },
        "/api/orders": {
            "get": {
                "summary": "Get all orders",
                "description": "Retrieve a list of all orders",
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Order"
                            }
                        }
                    },
                    "500": { "description": "Server error" }
                }
            },
            "post": {
                "summary": "Create a new order",
                "description": "Place a new order",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                ],
                "responses": {
                    "201": { "description": "Order placed successfully" },
                    "400": { "description": "Invalid input data" },
                    "500": { "description": "Server error" }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "summary": "Get an order by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "ID of the order"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    },
                    "404": { "description": "Order not found" },
                    "500": { "description": "Server error" }
                }
            },
            "put": {
                "summary": "Update an order by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "ID of the order to update"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "Order updated successfully" },
                    "400": { "description": "Invalid input data" },
                    "404": { "description": "Order not found" },
                    "500": { "description": "Server error" }
                }
            },
            "delete": {
                "summary": "Delete an order by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "ID of the order to delete"
                    }
                ],
                "responses": {
                    "200": { "description": "Order deleted successfully" },
                    "404": { "description": "Order not found" },
                    "500": { "description": "Server error" }
                }
            }
        }
    },
    "definitions": {
        "Product": {
            "type": "object",
            "properties": {
                "name": { "type": "string" },
                "description": { "type": "string" },
                "price": { "type": "number" },
                "stock": { "type": "number" },
                "category": { "type": "string" },
                "supplier": { "type": "string" }
            }
        },
        "Order": {
            "type": "object",
            "properties": {
                "customerName": { "type": "string" },
                "products": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "productId": { "type": "string" },
                            "quantity": { "type": "number" }
                        }
                    }
                },
                "totalAmount": { "type": "number" },
                "status": { "type": "string", "enum": ["Pending", "Shipped", "Delivered"] }
            }
        }
    }
}